requirements.txt
pytest~=7.4.2
webdriver-manager==4.0.1
selenium==4.14.0
allure-pytest==2.9.45
